+++ romancal/assign_wcs/assign_wcs_step.py
@@ -27,7 +27,7 @@ class AssignWcsStep(RomanStep):
     """ Assign a gWCS object to a science image.
     """
 
-    reference_file_types = []
+    reference_file_types = ['distortion']
 
     def process(self, input):
         reference_file_names = {}
@@ -41,12 +41,12 @@ class AssignWcsStep(RomanStep):
         return result
 
 
-def load_wcs(dmodel, reference_files):
+def load_wcs(input_model, reference_files=None):
     """ Create a gWCS object and store it in ``Model.meta``.
 
     Parameters
     ----------
-    dmodel : `~roman_datamodels.datamodels.WfiImage`
+    input_model : `~roman_datamodels.datamodels.WfiImage`
         The exposure.
     reference_files : dict
         A dict {reftype: reference_file_name} containing all
@@ -54,25 +54,74 @@ def load_wcs(dmodel, reference_files):
 
     Returns
     -------
-    dmodel : `~roman_datamodels.ImageModel`
+    output_model : `~roman_datamodels.ImageModel`
         The input image file with attached gWCS object.
         The data is not modified.
     """
-    if reference_files:
+    output_model = input_model.copy()
+
+    if reference_files is not None:
         for ref_type, ref_file in reference_files.items():
             if ref_file not in ["N/A", ""]:
                 reference_files[ref_type] = ref_file
             else:
                 reference_files[ref_type] = None
+    else:
+        reference_files = {}
 
+    # Frames
     detector = cf.Frame2D(name='detector', axes_order=(0, 1), unit=(u.pix, u.pix))
+    v2v3 = cf.Frame2D(name='v2v3', axes_order=(0, 1), axes_names=('v2', 'v3'), unit=(u.arcsec, u.arcsec))
     world = cf.CelestialFrame(reference_frame=coord.ICRS(), name='world')
-    tel2sky = pointing.v23tosky(dmodel)
-    pipeline = [Step(detector, tel2sky),
+
+    # Transforms between frames
+    distortion = wfi_distortion(output_model, reference_files)
+    tel2sky = pointing.v23tosky(output_model)
+
+    pipeline = [Step(detector, distortion),
+                Step(v2v3, tel2sky),
                 Step(world, None)]
     wcs = WCS(pipeline)
     if wcs.bounding_box is None:
-        wcs.bounding_box = wcs_bbox_from_shape(dmodel.data.shape)
-    dmodel.meta['wcs'] = wcs
-    dmodel.meta.cal_step['assign_wcs'] = 'COMPLETE'
-    return dmodel
+        wcs.bounding_box = wcs_bbox_from_shape(output_model.data.shape)
+
+    output_model.meta['wcs'] = wcs
+    output_model.meta.cal_step['assign_wcs'] = 'COMPLETE'
+
+    return output_model
+
+def wfi_distortion(model, reference_files):
+    """
+    Create the "detector" to "v2v3" transform for WFI
+
+    Parameters
+    ----------
+    model : `~roman_datamodels.datamodels.WfiImage`
+        The data model for processing
+    reference_files : dict
+        A dict {reftype: reference_file_name} containing all
+        reference files that apply to this exposure.
+
+    Returns
+    -------
+    The transform model
+    """
+
+    dist = rdm.DistortionRefModel(reference_files['distortion'])
+    transform = dist.coordinate_distortion_transform
+
+    try:
+        bbox = transform.bounding_box
+    except NotImplementedError:
+        # Check if the transform in the reference file has a ``bounding_box``.
+        # If not set a ``bounding_box`` equal to the size of the image after
+        # assembling all distortion corrections.
+        bbox = None
+    dist.close()
+
+    if bbox is None:
+        transform.bounding_box = wcs_bbox_from_shape(model.data.shape)
+    else:
+        transform.bounding_box = bbox
+
+    return transform